Replace the wait_for functionality in lib/http
The old implementation is too inflexible and causes a lot
of headaches. With this new one we will only retry on certain
errors and exit early otherwise.
The new system is also much more configurable with the use of
urllib3 Retry helper so that each component can decide on their
own retry strategy.