Add a client for sending internal events
This is the client that should be used together with the SSE v2 server
Also updated the rabbitmq log handler to send using this client instead
The logger will now publish all logs, not only user logs.

Note that the client is written in go, exported as a shared C library
and then loaded into python. The client can be built using make build
and is stored in the bindings directory. To enable these bindings when
testing locally an environment variable called BINDINGS must be set
to the bindings directory.

The reason for deciding to use Go -> C -> Python instead of using
the python library that exists is because the Python library is
lackluster in my opinion. Error messages were unclear and it was
quite slow. It is also written with asyncio which causes it to infect
everything else and made the client a lot more difficult to create.
With Go we get a reliable, and fast, publisher and we can use it
for our Go code as well and since we are going the route (at some
point) in re-writing ETOS into Go I don't see this as a problem.

This solution is not ready and there are TODOs. I just want feedback
on the approach.